+++ default_www/backend/core/js/backend.js
@@ -108,7 +108,7 @@ jsBackend.balloons =
 		$('.balloon:visible').each(function()
 		{
 			// search linked element
-			var linkedElement = $('*[data-message-id='+ $(this).attr('id') +']');
+			var linkedElement = $('*[data-message-id='+ $(this).prop('id') +']');
 
 			// linked item found?
 			if(linkedElement != null)
@@ -128,8 +128,8 @@ jsBackend.balloons =
 		var clickedElement = $(this);
 
 		// get linked balloon
-		var id = clickedElement.data('message-id');
-
+		var id = clickedElement.data('messageId');
+		
 		// rel available?
 		if(id != '')
 		{
@@ -220,15 +220,15 @@ jsBackend.controls =
 
 					if($(this).is(':checked'))
 					{
-						field.removeClass('disabled').attr('disabled', '');
+						field.removeClass('disabled').prop('disabled', '');
 						field.focus();
 					}
 
-					else field.addClass('disabled').attr('disabled', 'disabled');
+					else field.addClass('disabled').prop('disabled', 'disabled');
 				});
 
-				if(checkbox.is(':checked')) dropdown.removeClass('disabled').attr('disabled', '');
-				else dropdown.addClass('disabled').attr('disabled', 'disabled');
+				if(checkbox.is(':checked')) dropdown.removeClass('disabled').prop('disabled', '');
+				else dropdown.addClass('disabled').prop('disabled', 'disabled');
 			}
 		});
 	},
@@ -252,15 +252,15 @@ jsBackend.controls =
 
 					if($(this).is(':checked'))
 					{
-						field.removeClass('disabled').attr('disabled', '');
+						field.removeClass('disabled').prop('disabled', '');
 						field.focus();
 					}
 
-					else field.addClass('disabled').attr('disabled', 'disabled');
+					else field.addClass('disabled').prop('disabled', 'disabled');
 				});
 
-				if(checkbox.is(':checked')) textField.removeClass('disabled').attr('disabled', '');
-				else textField.addClass('disabled').attr('disabled', 'disabled');
+				if(checkbox.is(':checked')) textField.removeClass('disabled').prop('disabled', '');
+				else textField.addClass('disabled').prop('disabled', 'disabled');
 			}
 		});
 	},
@@ -282,13 +282,13 @@ jsBackend.controls =
 					$this = $(this);
 					
 					// disable all
-					$this.parents('.radiobuttonFieldCombo:first').find('input:not([name='+ radiobutton.attr('name') +']), select, textarea').addClass('disabled').attr('disbaled', 'disabled');
+					$this.parents('.radiobuttonFieldCombo:first').find('input:not([name='+ radiobutton.prop('name') +']), select, textarea').addClass('disabled').prop('disbaled', 'disabled');
 					
 					// get fields
-					var fields = $this.parents('li').find('input:not([name='+ radiobutton.attr('name') +']), select, textarea')
+					var fields = $this.parents('li').find('input:not([name='+ radiobutton.prop('name') +']), select, textarea')
 
 					// enable
-					fields.removeClass('disabled').attr('disabled', '');
+					fields.removeClass('disabled').prop('disabled', '');
 					
 					// set focus
 					$(fields[0]).focus();
@@ -307,8 +307,8 @@ jsBackend.controls =
 		$('.askConfirmation').each(function()
 		{
 			// get id
-			var id = $(this).data('message-id');
-			var url = $(this).attr('href');
+			var id = $(this).data('messageId');
+			var url = $(this).prop('href');
 
 			if(id != '' && url != '')
 			{
@@ -353,13 +353,13 @@ jsBackend.controls =
 			evt.preventDefault();
 
 			// get id
-			var id = $(this).data('message-id');
+			var id = $(this).data('messageId');
 
 			// bind
 			if(id != '')
 			{
 				// set target
-				$('#'+ id).data('message-id', $(this).attr('href'));
+				$('#'+ id).data('messageId', $(this).prop('href'));
 
 				// open dialog
 				$('#'+ id).dialog('open');
@@ -380,7 +380,7 @@ jsBackend.controls =
 			evt.stopPropagation();
 
 			// get id
-			var id = $(this).attr('href');
+			var id = $(this).prop('href');
 
 			if($(id).is(':visible'))
 			{
@@ -469,8 +469,8 @@ jsBackend.controls =
 	bindMassAction: function()
 	{
 		// set disabled
-		$('.tableOptions .massAction select').addClass('disabled').attr('disabled', 'disabled');
-		$('.tableOptions .massAction .submitButton').addClass('disabledButton').attr('disabled', 'disabled');
+		$('.tableOptions .massAction select').addClass('disabled').prop('disabled', 'disabled');
+		$('.tableOptions .massAction .submitButton').addClass('disabledButton').prop('disabled', 'disabled');
 
 		// hook change events
 		$('table input:checkbox').change(function(evt)
@@ -481,15 +481,15 @@ jsBackend.controls =
 			// any item checked?
 			if(table.find('input:checkbox:checked').length > 0)
 			{
-				table.find('.massAction select').removeClass('disabled').attr('disabled', '');
-				table.find('.massAction .submitButton').removeClass('disabledButton').attr('disabled', '');
+				table.find('.massAction select').removeClass('disabled').prop('disabled', '');
+				table.find('.massAction .submitButton').removeClass('disabledButton').prop('disabled', '');
 			}
 
 			// nothing checked
 			else
 			{
-				table.find('.massAction select').addClass('disabled').attr('disabled', 'disabled');
-				table.find('.massAction .submitButton').addClass('disabledButton').attr('disabled', 'disabled');
+				table.find('.massAction select').addClass('disabled').prop('disabled', 'disabled');
+				table.find('.massAction .submitButton').addClass('disabledButton').prop('disabled', 'disabled');
 			}
 		});
 
@@ -497,7 +497,7 @@ jsBackend.controls =
 		$('.tableOptions .massAction option').each(function()
 		{
 			// get id
-			var id = $(this).data('message-id');
+			var id = $(this).data('messageId');
 
 			if(typeof id != 'undefined')
 			{
@@ -515,7 +515,7 @@ jsBackend.controls =
 							$(this).dialog('close');
 
 							// submit the form
-							$($('select:visible option[data-message-id='+ $(this).attr('id') +']').parents('form')[0]).submit();
+							$($('select:visible option[data-message-id='+ $(this).prop('id') +']').parents('form')[0]).submit();
 						},
 						'{$lblCancel|ucfirst}': function()
 						{
@@ -547,10 +547,10 @@ jsBackend.controls =
 					var element = $(this).parents('.massAction').find('select[name=action] option:selected');
 
 					// if the rel-attribute exists we should show the dialog
-					if(typeof element.data('message-id') != 'undefined')
+					if(typeof element.data('messageId') != 'undefined')
 					{
 						// get id
-						var id = element.data('message-id');
+						var id = element.data('messageId');
 
 						// open dialog
 						$('#'+ id).dialog('open');
@@ -574,7 +574,7 @@ jsBackend.controls =
 		$('th .checkboxHolder input:checkbox').bind('change', function(evt)
 		{
 			// check or uncheck all the checkboxes in this datagrid
-			$($(this).closest('table').find('td input:checkbox')).attr('checked', $(this).is(':checked'));
+			$($(this).closest('table').find('td input:checkbox')).prop('checked', $(this).is(':checked'));
 
 			// set selected class
 			if($(this).is(':checked')) $($(this).parents().filter('table')[0]).find('tbody tr').addClass('selected');
@@ -587,11 +587,11 @@ jsBackend.controls =
 			// check mass checkbox
 			if($(this).closest('table').find('td.checkbox input:checkbox').length == $(this).closest('table').find('td.checkbox input:checkbox:checked').length)
 			{
-				$(this).closest('table').find('th .checkboxHolder input:checkbox').attr('checked', 'checked');
+				$(this).closest('table').find('th .checkboxHolder input:checkbox').prop('checked', 'checked');
 			}
 
 			// uncheck mass checkbox
-			else{ $(this).closest('table').find('th .checkboxHolder input:checkbox').attr('checked', ''); }
+			else{ $(this).closest('table').find('th .checkboxHolder input:checkbox').prop('checked', ''); }
 		});
 	},
 
@@ -622,7 +622,7 @@ jsBackend.controls =
 			{
 				// grab id
 				var id = $(this).data('id');
-				var wrapperId = $(this).attr('id');
+				var wrapperId = $(this).prop('id');
 
 				// hide all
 				$('#'+ wrapperId +' p.strength').hide();
@@ -707,7 +707,7 @@ jsBackend.controls =
 			evt.preventDefault();
 
 			// get id
-			var id = $(this).attr('href');
+			var id = $(this).prop('href');
 
 			// show/hide
 			$(id).toggle();
@@ -737,7 +737,7 @@ jsBackend.controls =
 	// bind target blank
 	bindTargetBlank: function()
 	{
-		$('a.targetBlank').attr('target', '_blank');
+		$('a.targetBlank').prop('target', '_blank');
 	},
 
 
@@ -891,7 +891,7 @@ jsBackend.forms =
 			// set options
 			$(this).datepicker('option', { 
 				dateFormat: data.mask, 
-				firstDay: data.firstday,
+				firstDate: data.firstday
 			}).datepicker('setDate', value);
 		});
 
@@ -966,7 +966,7 @@ jsBackend.forms =
 				var input = $(this);
 
 				// only do something when the current value and the placeholder are the same
-				if(input.val() == input.attr('placeholder'))
+				if(input.val() == input.prop('placeholder'))
 				{
 					// clear
 					input.val('');
@@ -982,10 +982,10 @@ jsBackend.forms =
 				var input = $(this);
 
 				// only do something when the input is empty or the value is the same as the placeholder
-				if(input.val() == '' || input.val() == input.attr('placeholder'))
+				if(input.val() == '' || input.val() == input.prop('placeholder'))
 				{
 					// set placeholder
-					input.val(input.attr('placeholder'));
+					input.val(input.prop('placeholder'));
 
 					// add class
 					input.addClass('placeholder');
@@ -1005,7 +1005,7 @@ jsBackend.forms =
 					var input = $(this);
 
 					// if the value and the placeholder are the same reset the value
-					if(input.val() == input.attr('placeholder')) input.val('');
+					if(input.val() == input.prop('placeholder')) input.val('');
 				});
 			});
 		}
@@ -1023,7 +1023,7 @@ jsBackend.forms =
 			$('form.submitWithLink').each(function()
 			{
 				// get id
-				var formId = $(this).attr('id');
+				var formId = $(this).prop('id');
 				var dontSubmit = false;
 
 				// validate id
@@ -1032,7 +1032,7 @@ jsBackend.forms =
 					// loop every button to be replaced
 					$('form#'+ formId + '.submitWithLink input:submit').each(function()
 					{
-						$(this).after(replaceHTML.replace('{label}', $(this).val()).replace('{id}', $(this).attr('id')).replace('{class}', 'submitButton button ' + $(this).attr('class'))).css({position:'absolute', top:'-9000px', left: '-9000px'}).attr('tabindex', -1);
+						$(this).after(replaceHTML.replace('{label}', $(this).val()).replace('{id}', $(this).prop('id')).replace('{class}', 'submitButton button ' + $(this).prop('class'))).css({position:'absolute', top:'-9000px', left: '-9000px'}).prop('tabindex', -1);
 					});
 
 					// add onclick event for button (button can't have the name submit)
@@ -1041,7 +1041,7 @@ jsBackend.forms =
 						evt.preventDefault();
 
 						// is the button disabled?
-						if($(this).attr('disabled') == 'disabled') return false;
+						if($(this).prop('disabled') == 'disabled') return false;
 						else $('form#'+ formId).submit();
 					});
 
@@ -1335,7 +1335,7 @@ jsBackend.tabs =
 			$('.tabs .ui-tabs-panel').each(function()
 			{
 				if($(this).find('.formError:visible').length > 0) {
-					$($('.ui-tabs-nav a[href="#'+ $(this).attr('id') +'"]').parent()).addClass('ui-state-error');
+					$($('.ui-tabs-nav a[href="#'+ $(this).prop('id') +'"]').parent()).addClass('ui-state-error');
 				}
 			});
 		}
@@ -1370,7 +1370,7 @@ jsBackend.tabs =
 			{
 				// prevent default
 				evt.preventDefault();
-				$('.tabs').tabs('select', $(this).attr('href'));
+				$('.tabs').tabs('select', $(this).prop('href'));
 			});
 		}
 	},
@@ -1397,7 +1397,7 @@ jsBackend.tinyMCE =
 		$('.clickToEdit').live('click', function(evt)
 		{
 			// get id
-			var id = $(this).siblings('textarea.inputEditor:first').attr('id');
+			var id = $(this).siblings('textarea.inputEditor:first').prop('id');
 
 			// validate id
 			if(typeof id != undefined)
@@ -1436,7 +1436,7 @@ jsBackend.tinyMCE =
 		var $tmp = $('<div />').html(object.content);
 
 		// replace target="_blank" with class="targetBlank"
-		$tmp.find('a.targetBlank').removeClass('targetBlank').attr('target', '_blank');
+		$tmp.find('a.targetBlank').removeClass('targetBlank').prop('target', '_blank');
 
 		// resave
 		object.content = utils.string.xhtml($tmp.html());