+++ default_www/frontend/core/js/frontend.js
@@ -4,12 +4,16 @@ if(!jsFrontend) { var jsFrontend = new Object(); }
 /**
  * Frontend related objects
  *
- * @author	Tijs Verkoyen <tijs@netlash.com>
+ * @author	Tijs Verkoyen <tijs@sumocoders.be>
  */
 jsFrontend =
 {
 	// datamembers
 	debug: false,
+	current:
+	{
+		language: '{$FRONTEND_LANGUAGE}'
+	},
 
 
 	// init, something like a constructor
@@ -57,7 +61,7 @@ jsFrontend.controls =
 	// bind target blank
 	bindTargetBlank: function()
 	{
-		$('a.targetBlank').attr('target', '_blank');
+		$('a.targetBlank').prop('target', '_blank');
 	},
 
 
@@ -106,7 +110,7 @@ jsFrontend.forms =
 			// set options
 			$(this).datepicker('option', { 
 				dateFormat: data.mask, 
-				firstDay: data.firstday,
+				firstDay: data.firstday
 			}).datepicker('setDate', value);
 		});
 
@@ -174,7 +178,7 @@ jsFrontend.forms =
 				var input = $(this);
 
 				// only do something when the current value and the placeholder are the same
-				if(input.val() == input.attr('placeholder'))
+				if(input.val() == input.prop('placeholder'))
 				{
 					// clear
 					input.val('');
@@ -190,10 +194,10 @@ jsFrontend.forms =
 				var input = $(this);
 
 				// only do something when the input is empty or the value is the same as the placeholder
-				if(input.val() == '' || input.val() == input.attr('placeholder'))
+				if(input.val() == '' || input.val() == input.prop('placeholder'))
 				{
 					// set placeholder
-					input.val(input.attr('placeholder'));
+					input.val(input.prop('placeholder'));
 
 					// add class
 					input.addClass('placeholder');
@@ -213,7 +217,7 @@ jsFrontend.forms =
 					var input = $(this);
 
 					// if the value and the placeholder are the same reset the value
-					if(input.val() == input.attr('placeholder')) input.val('');
+					if(input.val() == input.prop('placeholder')) input.val('');
 				});
 			});
 		}
@@ -233,14 +237,14 @@ jsFrontend.gravatar =
 		$('.replaceWithGravatar').each(function()
 		{
 			var element = $(this);
-			var gravatarId = element.data('gravatar-id');
-			var size = element.attr('height');
+			var gravatarId = element.data('gravatarId');
+			var size = element.prop('height');
 
 			// valid gravatar id
 			if(gravatarId != '')
 			{
 				// build url
-				var url = 'http://www.gravatar.com/avatar/'+ gravatarId + '?r=g&d=404';
+				var url = 'http://www.gravatar.com/avatar/' + gravatarId + '?r=g&d=404';
 
 				// add size if set before
 				if(size != '') url += '&s=' + size;
@@ -252,7 +256,7 @@ jsFrontend.gravatar =
 				// reset src
 				gravatar.onload = function()
 				{
-					element.attr('src', url).addClass('gravatarLoaded');
+					element.prop('src', url).addClass('gravatarLoaded');
 				}
 			}
 		});
@@ -274,22 +278,19 @@ jsFrontend.search =
 	// init, something like a constructor
 	init: function()
 	{
-		// split url to build the ajax-url
-		var chunks = document.location.pathname.split('/');
-
 		// autosuggest (search widget)
-		if($('input.autoSuggest').length > 0) jsFrontend.search.autosuggest(chunks[1], 55);
+		if($('input.autoSuggest').length > 0) jsFrontend.search.autosuggest(55);
 
 		// autocomplete (search results page: autocomplete based on known search terms)
-		if($('input.autoComplete').length > 0) jsFrontend.search.autocomplete(chunks[1]);
+		if($('input.autoComplete').length > 0) jsFrontend.search.autocomplete();
 
 		// livesuggest (search results page: live feed of matches)
-		if($('input.liveSuggest').length > 0 && $('#searchContainer').length > 0) jsFrontend.search.livesuggest(chunks[1]);
+		if($('input.liveSuggest').length > 0 && $('#searchContainer').length > 0) jsFrontend.search.livesuggest();
 	},
 
 
 	// autocomplete (search results page: autocomplete based on known search terms)
-	autocomplete: function(language)
+	autocomplete: function()
 	{
 		// autocomplete (based on saved search terms) on results page
 		$('input.autoComplete').autocomplete(
@@ -300,9 +301,9 @@ jsFrontend.search =
 				// ajax call!
 				$.ajax(
 				{
-					url: '/frontend/ajax.php?module=search&action=autocomplete',
+					url: '/frontend/ajax.php?module=search&action=autocomplete&language=' + jsFrontend.current.language,
 					type: 'GET',
-					data: 'term='+ request.term +'&language='+ language,
+					data: 'term=' + request.term,
 					success: function(data, textStatus)
 					{
 						// init var
@@ -335,14 +336,14 @@ jsFrontend.search =
 			{
 				url: '/frontend/ajax.php?module=search&action=save',
 				type: 'GET',
-				data: 'term='+ $(this).val() +'&language='+ language
+				data: 'term=' + $(this).val() + '&language=' + jsFrontend.current.language
 			});
 		});
 	},
 
 
 	// autosuggest (search widget)
-	autosuggest: function(language, length)
+	autosuggest: function(length)
 	{
 		// set default values
 		if(typeof length == 'undefined') length = 100;
@@ -356,9 +357,9 @@ jsFrontend.search =
 				// ajax call!
 				$.ajax(
 				{
-					url: '/frontend/ajax.php?module=search&action=autosuggest',
+					url: '/frontend/ajax.php?module=search&action=autosuggest&language=' + jsFrontend.current.language,
 					type: 'GET',
-					data: 'term='+ request.term +'&language='+ language +'&length='+ length,
+					data: 'term=' + request.term + '&length=' + length,
 					success: function(data, textStatus)
 					{
 						// init var
@@ -391,7 +392,7 @@ jsFrontend.search =
 			{
 				url: '/frontend/ajax.php?module=search&action=save',
 				type: 'GET',
-				data: 'term='+ $(this).val() +'&language='+ language
+				data: 'term=' + $(this).val() + '&language=' + jsFrontend.current.language
 			});
 		})
 		// and also: alter the autocomplete style: add description!
@@ -399,14 +400,14 @@ jsFrontend.search =
 		{
 			return $('<li></li>')
 			.data('item.autocomplete', item)
-			.append('<a><strong>'+ item.label +'</strong><br \>'+ item.desc +'</a>' )
+			.append('<a><strong>' + item.label + '</strong><br \>' + item.desc + '</a>' )
 			.appendTo(ul);
 		};
 	},
 
 
 	// livesuggest (search results page: live feed of matches)
-	livesuggest: function(language)
+	livesuggest: function()
 	{
 		// check if calls for live suggest are allowed
 		var allowCall = true;
@@ -426,9 +427,9 @@ jsFrontend.search =
 				// ajax call!
 				$.ajax(
 				{
-					url: '/frontend/ajax.php?module=search&action=livesuggest',
+					url: '/frontend/ajax.php?module=search&action=livesuggest&language=' + jsFrontend.current.language,
 					type: 'GET',
-					data: 'term='+ $(this).val() +'&language='+ language,
+					data: 'term=' + $(this).val(),
 					success: function(data, textStatus)
 					{
 						// allow for new calls