+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-2.2.0 (xxxx-xx-xx)
+2.3.0 (xxxx-xx-xx)
+--
+Bugfixes:
+
+* Core: when the metafields are disabled we don't have any values in the POST. When an error occurs in the other fields of the form the meta-fields would be cleared. As reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/164.
+* Pages: moving pages for a non-active language failed, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/163.
+
+
+Improvements:
+
+* Upgraded to jQuery 1.6.1
+* Upgraded to jQuery UI 1.8.13
+* Upgraded TinyMCE to 3.4.2
+
+2.2.0 (2011-06-01)
 --
 Bugfixes: